requirements.txt
pylint
requests
absl-py
websocket-client==0.56
numpy==1.16.2
gunicorn==20.0.4
git+git://github.com/Lu-Yi-Hsun/iqoptionapi
